requirements.txt
# local package
-e .

# external requirements

flake8
python-dotenv>=0.5.1
langchain-groq
langchain_openai==0.2.8
langchain==0.3.7
langchain_community==0.3.7
pymysql==1.1.1
chromadb==0.5.18
langchain-huggingface
langchain-groq
pandas
streamlit
langchain-chroma==0.1.4
faiss-cpu

# backwards compatibility
# pathlib2
